js: split nft bindings
